refactor: CLLoaction 뷰컨에서 분리
